clear refactoring
