+++ templete/first_page.html
@@ -5,7 +5,6 @@
     <link href="/templete/first_page.css" rel="stylesheet"/>
     <title>Document</title>
     <script src="/javascript/utils.js"></script>
-    <script src="/javascript/all_share.js"></script>
     <script src="/javascript/move.js"></script>
 </head>
 <body>
