+++ templete/today.html
@@ -1,12 +1,15 @@
 <!DOCTYPE html>
 <html>
+
 <head>
     <link href="https://fonts.googleapis.com/css?family=Roboto&display=swap" rel="stylesheet" />
     <link href="/templete/today.css" rel="stylesheet" />
     <title>Document</title>
-    <script type="module" src="/javascript/today.js"></script>
 </head>
-<body>
+<script src="/javascript/utils.js"></script>
+<script src="/javascript/today.js"></script>
+
+<body onload="init_loader()">
     <div class="today">
         <button class="web_name" onclick="location.href='first_page'">편입영어</button>
         <span class="current_date"></span>
