+++ templete/total.html
@@ -5,7 +5,8 @@
     <link href="https://fonts.googleapis.com/css?family=Roboto&display=swap" rel="stylesheet"/>
     <link href="/templete/total.css" rel="stylesheet"/>
     <title>Document</title>
-    <script type="module" src="/javascript/total.js"></script>
+    <script src="/javascript/utils.js"></script>
+    <script src="/javascript/total.js"></script>
 </head>
 <body>
     <div class="total">
