+++ templete/total_review.html
@@ -4,9 +4,10 @@
     <link href="https://fonts.googleapis.com/css?family=Roboto&display=swap" rel="stylesheet" />
     <link href="/templete/total_review.css" rel="stylesheet" />
     <title>Document</title>
-    <script type="module" src="/javascript/total_review.js"></script>
+    <script src="/javascript/utils.js"></script>
+    <script src="/javascript/total_review.js"></script>
 </head>
-<body>
+<body onload="first_load()">
 <div class="total_review">
     <form method="GET">
         <button class="web_name" onclick="reload_alert()">편입영어</button>
@@ -16,9 +17,11 @@
     <button class="left_button" onclick="left_click()"></button>
     <button class="right_button" onclick="right_click()"></button>
     <div class="word_card"></div>
-    <p class="english"></p>
-    <p class="meaning"></p>
-    <p class="sentence"></p>
+    <div class="word_information">
+        <p class="english"></p>
+        <p class="meaning"></p>
+        <p class="sentence"></p>
+    </div>
     <img class="O_button" src="http://127.0.0.1:8000/o_button" onclick="left_click()"></img>
     <img class="X_button" src="http://127.0.0.1:8000/x_button" onclick="right_click()"></img>
     <img class="return_button" src="http://127.0.0.1:8000/return_button"></img>
